+++ weady/weady/UserFavoriteLocation/Services/UserFavoriteLocationServices.swift
@@ -5,6 +5,7 @@
 //  Created by Yoonseo on 7/31/25.
 //
 
+/*
 import Foundation
 
 final class UserFavoriteLocationServices {
@@ -63,3 +64,5 @@ final class UserFavoriteLocationServices {
     }
 
 }
+
+*/
