+++ weady/weady/weadyApp.swift
@@ -11,7 +11,9 @@ import SwiftUI
 struct weadyApp: App {
     var body: some Scene {
         WindowGroup {
-            TermsAgreementView()
+         //TermsAgreementView()
+        SettingView()
+       // NotificationView()
         }
     }
 }
